% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAggrMetrics.R
\name{calculateAggrMetrics}
\alias{calculateAggrMetrics}
\title{Calculate aggregate metrics for each route.}
\usage{
calculateAggrMetrics(df, pathToBRouter, profile = "trekking")
}
\arguments{
\item{df}{A dataframe or datatable containing all origin-destination pairs. The table has to contain the columns 'startLat', 'startLon', 'endLat', 'endLon' in WGS84 coordinates}

\item{pathToBrouter}{Path to installation folder of brouter}
}
\value{

}
\description{
Single core version for calculating aggregate metrics for routes based on calculateRoute. The output is a dataframe containing the travel time, distance, total elevation gain, upward and downward slope and energy expenditure for each route.
Optionally, the dataframe can contain the columns 'bikerPower', 'totalMass', 'dragCoefficient', 'rollingResistance' and 'maxSpeed'. In this case these parameters will be passed to the route calculation. If these columns are not available, default values will be used.
If brouter is not able to calculate the route, all variables will be returned with the value -99. Reasons for this may be coordinates outside the downloaded segment, or too far away from any nearby road.
}
\examples{


}
