% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAggrMetricsForParallel_OLDVERSION.R
\name{calculateAggrMetricsForParallelOLDVERSION}
\alias{calculateAggrMetricsForParallelOLDVERSION}
\title{Calculate aggregate metrics for each route.}
\usage{
calculateAggrMetricsForParallelOLDVERSION(
  matrix = NULL,
  pathToBrouter = NULL,
  profile = "trekking"
)
}
\arguments{
\item{pathToBrouter}{Path to installation folder of brouter}

\item{profile}{The routing profile}

\item{line}{The line of the matrix passed}
}
\value{

}
\description{
The output is a dataframe containing the travel time, distance, total elevation gain, upward and downward slope and energy expenditure for each route.
If brouter is not able to calculate the route, all variables will be returned with the value -99. Reasons for this may be coordinates outside the downloaded segment, or too far away from any nearby road.
}
\examples{

}
\keyword{internal}
