% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRoute.R
\name{calculateRoute}
\alias{calculateRoute}
\title{Calculate route between two points}
\usage{
calculateRoute(
  startLat,
  startLon,
  endLat,
  endLon,
  bikerPower = 100,
  totalMass = 90,
  dragCoefficient = 0.559,
  rollingResistance = 0.0077,
  maxSpeed = 45,
  profile = "trekking",
  outputFormat = "csv",
  serverNodeId = 1
)
}
\arguments{
\item{startLat}{Latitude of start location}

\item{startLon}{Longitude of start location}

\item{endLat}{Latitude of end location}

\item{endLon}{Longitude of end location}

\item{bikerPower}{The total average power put on the pedals in Watts. Defaults to 100W}

\item{totalMass}{The total weight of biker, bike and cargo in kg. Defaults to 90kg}

\item{dragCoefficient}{The wind drag coefficient in m2. Defaults to 0.559}

\item{rollingResistance}{The rolling resistance of the underground. Default value assumes dry asphalt, 0.0077}

\item{maxSpeed}{The maximum speed achieved by the bike in km/h. Defaults to 45 km/h.}

\item{profile}{Routing profile. Defaults to "trekking"}

\item{outputFormat}{one of "csv", "gpx" or "geojson". Defaults to "csv"}

\item{serverNodeId}{The node where the server is running on. Defaults to 1. Do not change for single core usage.}
}
\value{
Either a dataframe of the track (outputFormat="csv"), or a SpatialPointsDataFrame (outputFormat="gpx"). In the SpatialPointsDataFrame, elevation data of each point is contained in the data.
}
\description{
The route between two points is calculated for the given coordinates. The default values for rolling resistance and drag coefficient
correspond to values found by Tengatti and Bigazzi (2018): Physical characteristics and resistance parameters of typical urban cyclists.
brouterR ignores the bikerPower, totalMass, dragCoefficient, rollingResistance and maxSpeed available in the brouter routing profiles and uses the values provided via R instead.
This functions needs the server to be set manually before running it with the setServers and then the startServers functions.
}
\examples{


}
