% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installbrouterR.R
\name{installbrouterR}
\alias{installbrouterR}
\title{Install the brouterR files to a local path folder}
\usage{
installbrouterR(installationFolderPath = NULL)
}
\arguments{
\item{installationFolderPath}{installation path}
}
\value{

}
\description{
Install the brouterR files to a local path folder
}
