% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBasedRouteCalculator.R
\name{listBasedRouteCalculator}
\alias{listBasedRouteCalculator}
\title{Calculate loops based on input as an element of a list.}
\usage{
listBasedRouteCalculator(elem = NULL, profile = "trekking")
}
\arguments{
\item{elem}{An element of a list}

\item{profile}{The routing profile to be used}
}
\value{

}
\description{
To be used with lapply!
}
\examples{

}
\keyword{internal}
