% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleBRoutes.R
\name{multipleBRoutes}
\alias{multipleBRoutes}
\title{Parallel routes calculation}
\usage{
multipleBRoutes(
  preset = NULL,
  df = NULL,
  nrOfNodes = 1,
  pathToBrouter = NULL,
  profile = NULL,
  optional = NULL
)
}
\arguments{
\item{df}{A dataframe or datatable containing all origin-destination pairs. The table has to contain the columns 'id', 'startLat', 'startLon', 'endLat', 'endLon' in WGS84 coordinates}

\item{nrOfNodes}{Number of nodes the router is supposed to run on. Suggested to always use at max 1 core less than the total amount of cores in processor.}

\item{pathToBrouter}{Path to installation folder of brouter.}

\item{profile}{The brouter profile to be used in the routing.}

\item{optional}{Optional arguments: A list containing a combination of column names which have to exactly match the following names: 'bikerPower', 'totalMass', 'dragCoefficient', 'rollingResistance', 'maxSpeed'}
}
\value{
A dataframe containing the route 'id', 'travelTime', 'distance', 'energy, 'avgSlopeUp, 'avgSlopeDown'
}
\description{
Calculate routes with parallel routine for improved speed. The output is a dataframe containing the travel times, distance, total elevation gain, upward and downward slope and energy expenditure for the route
The calculated values are a summary of the detailed csv-file that brouter outputs at a route request. The original file can be called with the calculateRoute function
or visualized online at http://www.brouter.de/brouter-web by clicking on the Data option after calculating a route.
}
\examples{

}
