% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRoute.R
\name{plotRoute}
\alias{plotRoute}
\title{Plot html map view of route}
\usage{
plotRoute(
  route = NULL,
  plotOnViewer = TRUE,
  backgroundTile = providers$OpenStreetMap,
  tileOpacity = 0.8,
  routeOpacity = 0.8,
  routeColour = "blue",
  routeWeight = 2
)
}
\arguments{
\item{route}{A route result from calculateRoute as an st_linestring object.}

\item{plotOnViewer}{If the map should directly be plotted on RStudio-Viewer or not. If TRUE, nothing is returned, if FALSE, the function returns an object, which can be saved to file using the saveWidget function of the htmlwidgets package.}

\item{backgroundTile}{A background tile of your choice (see list of options in leaflet::providers$)}

\item{tileOpacity}{Opacity of background tiles.}

\item{routeOpacity}{Opacity of displayed route.}

\item{routeColour}{Colour of displayed route.}

\item{routeWeight}{Weight of displayed route.}
}
\value{

}
\description{
Uses the leaflet package to plot the route on a map.
}
\examples{


}
