% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setServers.R
\name{setServers}
\alias{setServers}
\title{Set up servers for running brouter}
\usage{
setServers(nrOfNodes = 1, nodeRAM = 128, pathToBRouter = NULL)
}
\arguments{
\item{nrOfNodes}{Define number of parallel nodes. Defaults to 1}

\item{nodeRAM}{Define RAM allocated to each node. Defaults to 128mb}

\item{pathToBRouter}{Set the path to brouter directory with correct folder structure.}
}
\value{

}
\description{
Set up servers for running brouter
}
\examples{

}
\keyword{internal}
