% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitForCores.R
\name{splitForCores}
\alias{splitForCores}
\title{Split df into list of lists to be passed to each core}
\usage{
splitForCores(df, nrOfNodes)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Split df into list of lists to be passed to each core
}
\keyword{internal}
