% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startServers.R
\name{startServers}
\alias{startServers}
\title{Start running servers}
\usage{
startServers(pathToBRouter = NULL, noServers = 1)
}
\arguments{
\item{noServers}{number of servers, defaults to 1}

\item{pathToBrouterDir}{path to Brouter directory}
}
\value{
none
}
\description{
Start running servers
}
\examples{

}
